changed the name on select entities, added commas to health and stored the static part of the name.
Make an item display (Only specified NBT is CustomName) ride every skeleton/husk/wither_skeleton that has the visible CustomName NBT instead of the skeleton. All other data like the raw name is still stored in the skeleton's head item. Only the CustomName NBT should be in the item display instead of the skeleton
Add commas in health and maxhealth (For the name)
Store everything after the current health directly in ArmorItems[3].tag.HealthBarName ⁠https://discord.com/channels/1031998050861858907/1031999849094189056/1135260053503492236 (With the ' around the NBT. Basically the CustomName NBT can be copied over, just without the current health, and with the ' around the nbt)