fixed mob triggers and reordered the inputs.
I noticed that the ,RandomAttack:[{ID:,Timer:}] part is always there even if there's no input. And the abilities (RandomAttack, WhenDamaged, WhenKilled, WhenPoofed, WhenTargeting) should be in Triggers:{} NBT, so it's grouped together nicely and easier to see what they do

One last thing: Would it be possible to move the abilities down (Below the knockback resistance) and signify that they're abilities? Like a title or anything that seperates it from the rest.